// frontend-v2/src/app/core/api/models/cliente-input.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ClienteEnderecoInput } from './cliente-endereco-input';

export interface ClienteInput {
  nome: string;
  tel_1?: string;
  tel_2?: string;
  tel_3?: string;
  documento?: string;
  enderecos: Array<ClienteEnderecoInput>;
}